lint: small cleanup
