+++ src/server/passportconfig.ts
@@ -1,6 +1,6 @@
 import passport from "passport";
 import { Strategy as GoogleStrategy } from "passport-google-oauth20";
-import { prisma } from "./store";
+import { prisma } from "./db";
 
 passport.use(
   new GoogleStrategy(
